+++ crates/rust-client/src/store/sqlite_store/store.sql
@@ -8,18 +8,28 @@ CREATE TABLE settings (
     CONSTRAINT settings_name_is_not_empty CHECK (length(name) > 0)
 ) STRICT, WITHOUT ROWID;
 
--- Create account_code table
-CREATE TABLE account_code (
-    root TEXT NOT NULL,         -- root of the Merkle tree for all exported procedures in account module.
-    code BLOB NOT NULL,         -- serialized account code.
+
+-- Create account_procedures table
+CREATE TABLE account_procedures (
+    mast_forest_root TEXT NOT NULL, -- root of the Merkle tree for all exported procedures in account module.
+    procedure_root TEXT NOT NULL,   -- mast root of the individual procedure.
+    FOREIGN KEY (mast_forest_root) REFERENCES mast_forests(root),
+    PRIMARY KEY (mast_forest_root, procedure_root)
+);
+
+-- Create mast_forests table
+CREATE TABLE mast_forests (
+    root TEXT NOT NULL,             -- root of the Merkle tree for all exported procedures in account module.
+    procedure_info BLOB NOT NULL,   -- serialized procedure info list.
+    mast BLOB NOT NULL,             -- serialized mast forest.
     PRIMARY KEY (root)
 );
 
 -- Create account_storage table
 CREATE TABLE account_storage (
-    root TEXT NOT NULL,         -- root of the account storage Merkle tree.
+    commitment TEXT NOT NULL,   -- commitment of the account storage Merkle tree.
     slots BLOB NOT NULL,        -- serialized key-value pair of non-empty account slots.
-    PRIMARY KEY (root)
+    PRIMARY KEY (commitment)
 );
 
 -- Create account_vaults table
@@ -32,25 +42,25 @@ CREATE TABLE account_vaults (
 -- Create foreign_account_code table
 CREATE TABLE foreign_account_code(
     account_id TEXT NOT NULL,              -- ID of the account
-    code_root TEXT NOT NULL,               -- Root of the account_code
+    code_root TEXT NOT NULL,               -- root of the Merkle tree for all exported procedures in account module.
     PRIMARY KEY (account_id),
-    FOREIGN KEY (code_root) REFERENCES account_code(root)
+    FOREIGN KEY (code_root) REFERENCES mast_forests(root)
 );
 
 -- Create accounts table
 CREATE TABLE accounts (
     account_commitment TEXT NOT NULL UNIQUE,    -- Account state commitment
     id UNSIGNED BIG INT NOT NULL,               -- Account ID.
-    code_root TEXT NOT NULL,                    -- Root of the account_code
-    storage_root TEXT NOT NULL,                 -- Root of the account_storage Merkle tree.
+    code_root TEXT NOT NULL,                    -- Root of the Merkle tree for all exported procedures in account module.
+    storage_commitment TEXT NOT NULL,           -- Commitment to the account_storage Merkle tree.
     vault_root TEXT NOT NULL,                   -- Root of the account_vault Merkle tree.
     nonce BIGINT NOT NULL,                      -- Account nonce.
     committed BOOLEAN NOT NULL,                 -- True if recorded, false if not.
     account_seed BLOB NULL,                     -- Account seed used to generate the ID. Expected to be NULL for non-new accounts
     locked BOOLEAN NOT NULL,                    -- True if the account is locked, false if not.
     PRIMARY KEY (account_commitment),
-    FOREIGN KEY (code_root) REFERENCES account_code(root),
-    FOREIGN KEY (storage_root) REFERENCES account_storage(root),
+    FOREIGN KEY (code_root) REFERENCES mast_forests(root),
+    FOREIGN KEY (storage_commitment) REFERENCES account_storage(commitment),
     FOREIGN KEY (vault_root) REFERENCES account_vaults(root)
 
     CONSTRAINT check_seed_nonzero CHECK (NOT (nonce = 0 AND account_seed IS NULL))