+++ crates/web-client/src/new_transactions.rs
@@ -26,6 +26,7 @@ impl WebClient {
         transaction_request: &TransactionRequest,
     ) -> Result<TransactionResult, JsValue> {
         self.fetch_and_cache_account_auth_by_account_id(account_id).await?;
+        self.fetch_and_cache_mast_forests().await?;
 
         if let Some(client) = self.get_mut_inner() {
             let native_transaction_execution_result: NativeTransactionResult = client